Make some arrays constexpr
Use Unicode literals, instead of using UTF-8 code units.

And change the key type from LONG to UINT, to match the type
of the value obtained from MapVirtualKeyW.

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184193
Tested-by: Jenkins

+++ vcl/inc/win/salframe.h
@@ -161,10 +161,9 @@ bool UseDarkMode();
 bool OSSupportsDarkMode();
 
 // get foreign key names
-namespace vcl_sal {
-    OUString getKeysReplacementName(
-        std::u16string_view pLang,
-        LONG nSymbol );
+namespace vcl_sal
+{
+OUString getKeysReplacementName(std::u16string_view pLang, UINT nSymbol);
 }
 
 /* vim:set shiftwidth=4 softtabstop=4 expandtab: */
